+++ scripts/specificSummons.js
@@ -19,6 +19,10 @@ import { incarnateDetails } from "./specificCases/incarnate.js";
 import { getEidolon } from "./specificClasses/summoner.js";
 import { isSummonSourceDisabled } from "./disableItems.js";
 import { getNecromancerProf } from "./specificClasses/necromancer.js";
+import {
+  dancingWeaponDialog,
+  getJB2aPath,
+} from "./specificCases/dancingWeapon.js";
 
 export async function getSpecificSummonDetails(
   uuid,
@@ -118,6 +122,9 @@ const getSummonHandlers = () => ({
     handlers.necromancer.handleRecurringNightmare,
   [SOURCES.NECROMANCER.SKELETAL_LANCERS]:
     handlers.necromancer.handleSkeletalLancers,
+
+  // Psychic
+  [SOURCES.PSYCHIC.DANCING_BLADE]: handlers.psychic.handleDancingBlade,
 
   // Summon
   [SOURCES.MISC.PHANTASMAL_MINION]: handlers.summon.handlePhantasmalMinion,
@@ -652,7 +659,6 @@ const handlers = {
       ];
     },
   },
-
   wall: {
     handleWallOfIce: async (data) => {
       const type = data.ignoreDialogue
@@ -821,7 +827,6 @@ const handlers = {
       ];
     },
   },
-
   necromancer: {
     handleBindHeroicSpiritStrike: (data) => {
       return [
@@ -932,6 +937,46 @@ const handlers = {
               rollOptions: data.summonerRollOptions,
             }),
           ],
+        },
+      ];
+    },
+  },
+  psychic: {
+    handleDancingBlade: async (data) => {
+      const summonerActor = game.actors.get(data.summonerActorId);
+      const isAmped = data?.summonerRollOptions?.includes("amp-spell");
+      let weapon = {
+        id: "",
+        name: "",
+        tokenArt: getJB2aPath(
+          "jb2a.spiritual_weapon.shortsword.01.spectral.02.green",
+        ),
+      };
+      let dialogResult = {};
+      let effect = {};
+      if (!data.ignoreDialogue) {
+        dialogResult = await dancingWeaponDialog(summonerActor, isAmped);
+      }
+
+      if (dialogResult.effect) {
+        effect = dialogResult.effect;
+      }
+      if (dialogResult.weapon) {
+        weapon = dialogResult.weapon;
+      }
+
+      return [
+        {
+          specific_uuids: [CREATURES.PSYCHIC.DANCING_BLADE],
+          rank: data.rank,
+          modifications: {
+            "system.details.level.value": data.rank,
+            "prototypeToken.texture.src": weapon.tokenArt,
+          },
+          itemsToAdd: [effect],
+          crosshairParameters: {
+            distance: canvas.grid.distance / 4,
+          },
         },
       ];
     },